requirements.txt
asttokens==2.4.1
attrs==23.2.0
backcall==0.2.0
beautifulsoup4==4.12.3
bleach==6.1.0
blinker==1.8.2
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
decorator==5.1.1
defusedxml==0.7.1
docopt==0.6.2
executing==2.0.1
fastjsonschema==2.19.1
Flask==3.0.3
Flask-MySQLdb==2.0.0
idna==3.7
install==1.3.5
ipython==8.12.3
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
mistune==3.0.2
mysql-connector-python==8.4.0
mysql-connector-python-rf==2.2.2
mysqlclient==2.2.4
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
packaging==24.0
pandocfilters==1.5.1
parso==0.8.4
pickleshare==0.7.5
pipreqs==0.5.0
platformdirs==4.2.2
prompt-toolkit==3.0.43
pure-eval==0.2.2
Pygments==2.18.0
python-dateutil==2.9.0.post0
pywin32==306
pyzmq==26.0.3
referencing==0.35.1
requests==2.32.2
rpds-py==0.18.1
six==1.16.0
soupsieve==2.5
stack-data==0.6.3
tabulate==0.9.0
tinycss2==1.3.0
tornado==6.4
traitlets==5.14.3
urllib3==2.2.1
vboxapi==1.0
wcwidth==0.2.13
webencodings==0.5.1
Werkzeug==3.0.3
yarg==0.1.9
